% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uptake_Plots.R
\name{AllRegressionPlots}
\alias{AllRegressionPlots}
\title{All regression plots for a full HDX/DHX experiment}
\usage{
AllRegressionPlots(Allcents, pepIDs = peptide.identifications, pepnumber,
  times)
}
\description{
Calculates the regression curves (linear, logarithmic and
polynomial) for a complete HDX/DHX experiment and saves each set of 3 plots
in a pdf. All pdfs are saved in a directory called: "AllRegressionPlots"
}
\examples{
AllRegressionPlots(Allcents,pepnumber,times)
}
\author{
Sam MacIntyre \email{smacintyre@csiro.au}
}
\keyword{linear,}
\keyword{logarithmic,}
\keyword{pdf,}
\keyword{plots}
\keyword{polynomial,}
\keyword{regression,}
