% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Export_function.R
\name{AvCent}
\alias{AvCent}
\title{Average Centroid Calculation}
\usage{
AvCent(AllpeptideSamples)
}
\arguments{
\item{AllpeptideSamples}{A singular mass spectrum}
}
\description{
Averages the centroids for each time point across each
technical replicate
}
\examples{
[a,b,envelope] <- widthFinder(peaks,spectrum,bpint = 0.5)
}
\author{
Sam MacIntyre \email{smacintyre@csiro.au}
}
\keyword{centroids,average,sample}
