% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uptake_Plots.R
\name{AverageCents}
\alias{AverageCents}
\title{Average centroids for a complete HDX/DHX experiment}
\usage{
AverageCents(PepDX = peptide.featuresD, PepHDX = peptide.features,
  SNR = 5, pepnums)
}
\description{
Calculates average centroids across each sample and each time
point
}
\examples{
[x,y] <- LinIntersect(LinCoefs)
}
\author{
Sam MacIntyre \email{smacintyre@csiro.au}
}
\keyword{intersection}
\keyword{linear,}
