% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculation_functions.R
\name{centroidCalc}
\alias{centroidCalc}
\title{Centroid calculation}
\usage{
centroidCalc(v, spectrum)
}
\arguments{
\item{v}{Vector of end points of peak width}

\item{spectrum}{Mass spectrum}
}
\description{
Calculates the centroid for a spectrum given the peak width
}
\examples{
c <- centroidCalc(v,spectrum)
}
\author{
Sam MacIntyre \email{smacintyre@csiro.au}
}
\keyword{centroids,width}
