% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_functions.R
\name{centroidPlot}
\alias{centroidPlot}
\title{Centroid plotting function}
\usage{
centroidPlot(cent, s)
}
\arguments{
\item{cent}{centroid value}

\item{s}{spectrum}
}
\description{
Given the value of the centroid and the spectrum, plots
the centroid
}
\author{
Sam MacIntyre \email{smacintyre@csiro.au}
}
\keyword{centroid,}
\keyword{plot,}
\keyword{segment}
