% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{import.identifications}
\alias{import.identifications}
\title{Imports Scaffold file for peptide identification}
\usage{
import.identifications(path = getwd(), filename = "Scaffold.txt")
}
\arguments{
\item{path}{A singular mass spectrum}

\item{ids}{}

\item{tol}{}
}
\description{
Imports the Scaffold.txt file to create the an identifications
file
}
\examples{
PepIDs <- ScaffoldImport()
}
\author{
Sam MacIntyre \email{smacintyre@csiro.au}
}
\keyword{Scaffold}
\keyword{identification,}
\keyword{import,}
\keyword{peptide,}
