% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{importNew}
\alias{importNew}
\title{Import Spectra function}
\usage{
importNew(path = getwd(), ids = peptide.identifications, tol = 0.006)
}
\arguments{
\item{path}{}

\item{ids}{}

\item{tol}{}
}
\description{
Imports mass spectra and creates a list of feature matrices 
using the MALDIquant package
}
\examples{
peptide.features <- importNew()
}
\author{
Sam MacIntyre \email{smacintyre@csiro.au}
}
\keyword{MALDIquant}
\keyword{feature,}
\keyword{import,}
\keyword{spectra,}
