% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculation_functions.R
\name{linearInterp}
\alias{linearInterp}
\title{Linear Interpolation Function}
\usage{
linearInterp(peaks)
}
\arguments{
\item{peaks}{list of mass peaks}
}
\description{
Given a set of mass peaks, finds the linear interpolation
between them
}
\examples{
interp <- linearInterp(peaks)
}
\author{
Sam MacIntyre \email{smacintyre@csiro.au}
}
\keyword{linear,interpolation,peaks,}
