% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculation_functions.R
\name{linearsolve}
\alias{linearsolve}
\title{Linear Equation Solver}
\usage{
linearsolve(p1, p2, value)
}
\arguments{
\item{p1}{Vector containing first point [x1,y1]}

\item{p2}{Vector containing second point [x2,y2]}

\item{value}{Scalar denoting required y value for intersection}
}
\description{
Finds the x coordinate of the intersection point of a linear
function with a specified value
}
\examples{
x <- linearsolve(p1,p2,value)
}
\author{
Sam MacIntyre \email{smacintyre@csiro.au}
}
\keyword{equation}
\keyword{interpolation,}
\keyword{linear,}
\keyword{solver,}
