% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uptake_Plots.R
\name{LinIntersect}
\alias{LinIntersect}
\title{Linear intersection}
\usage{
LinIntersect(LinCoefs)
}
\description{
Calculates the intersection point of two linear functions
given their coefficients
}
\examples{
[x,y] <- LinIntersect(LinCoefs)
}
\author{
Sam MacIntyre \email{smacintyre@csiro.au}
}
\keyword{intersection}
\keyword{linear,}
