% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uptake_Plots.R
\name{LogIntersect}
\alias{LogIntersect}
\title{Logarithmic intersection}
\usage{
LogIntersect(LogCoefs)
}
\description{
Calculates the intersection point of two logarithmic functions
given their coefficients
}
\examples{
[x,y] <- LogIntersect(LogCoefs)
}
\author{
Sam MacIntyre \email{smacintyre@csiro.au}
}
\keyword{intersection}
\keyword{log,logarithmic,}
