% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Export_function.R
\name{mainCentMerge}
\alias{mainCentMerge}
\title{Centroid calculation and uptakes for a complete peptide}
\usage{
mainCentMerge(samplematrix, charge = 1, SNR = 10, bpint = 0.5,
  sample = 0, time = 0)
}
\arguments{
\item{samplematrix}{Matrix of all timepoints and samples for a single peptide}

\item{SNR}{time point}

\item{bpint}{timepoints}

\item{sample}{masses}

\item{time}{base peak intensity}
}
\description{
Allows the calculation of centroids for each time point and
sample for a single peptide. Also prompts manual centroid calculation if
automatic value is not appropriate. Each centroid is saved as a pdf in the
current directory.
}
\examples{
Centroids <- mainCentMerge(Peptide1,SNR=5)
}
\author{
Sam MacIntyre \email{smacintyre@csiro.au}
}
\keyword{locator}
\keyword{manual,}
\keyword{peaks,}
