% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Export_function.R
\name{mainCentNewMod}
\alias{mainCentNewMod}
\title{Centroid relative uptakes calculation and export for a single peptide}
\usage{
mainCentNewMod(samplematrix, charge = 1, SNR = 10, bpint = 0.5,
  sample = 0, time = 0)
}
\arguments{
\item{bpint}{base peak intensity}

\item{spectrum}{A singular mass spectrum}

\item{i}{sample number}

\item{j}{time point}

\item{TimeP}{timepoints}

\item{mass}{masses}
}
\description{
Calculates the centroids for each sample and timepoint for
each
}
\examples{
[c,v] <- manualCentroids2(s,1,2,mass)
}
\author{
Sam MacIntyre \email{smacintyre@csiro.au}
}
\keyword{locator}
\keyword{manual,}
\keyword{peaks,}
