% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Export_function.R
\name{manualCentroids2}
\alias{manualCentroids2}
\title{Manual Peak Selection}
\usage{
manualCentroids2(spectrum, i, j, TimeP, mass, bpint = 0.5)
}
\arguments{
\item{spectrum}{A singular mass spectrum}

\item{i}{sample number}

\item{j}{time point}

\item{TimeP}{timepoints}

\item{mass}{masses}

\item{bpint}{base peak intensity}
}
\description{
Enables the locator to be used to select peaks and hence
calculates the centroid and peak width
}
\examples{
[c,v] <- manualCentroids2(s,1,2,mass)
}
\author{
Sam MacIntyre \email{smacintyre@csiro.au}
}
\keyword{locator}
\keyword{manual,}
\keyword{peaks,}
