% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculation_functions.R,
%   R/HXMacspressAnalysisUpdated1117.R
\name{peakPick}
\alias{peakPick}
\alias{peakPick}
\title{Peak Picking Function}
\usage{
peakPick(spectrum, SNR = 10)

peakPick(spectrum, SNR = 10)
}
\arguments{
\item{spectrum}{A singular mass spectrum}

\item{spectrum}{A singular mass spectrum}
}
\description{
Identifies peaks in the spectrum

Identifies peaks in the spectrum
}
\examples{
p <- peakPick(s)
p <- peakPick(s)
}
\author{
Sam MacIntyre \email{smacintyre@csiro.au}

Sam MacIntyre \email{smacintyre@csiro.au}
}
\keyword{SNR}
\keyword{peaks,}
