% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uptake_Plots.R
\name{PlotDHAll}
\alias{PlotDHAll}
\title{All regression plots for a single peptide}
\usage{
PlotDHAll(pepnum, pepIDS = peptide.identifications, HDXcents, DXcents, times)
}
\description{
Calculates the regression curves (linear, logarithmic and
polynomial) for a full peptide
}
\examples{
PlotDHAll(pepnum,DXcents,DXcents,times)
}
\author{
Sam MacIntyre \email{smacintyre@csiro.au}
}
\keyword{linear,}
\keyword{logarithmic,}
\keyword{pdf,}
\keyword{plots}
\keyword{polynomial,}
\keyword{regression,}
