% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uptake_Plots.R
\name{PlotDHLinear}
\alias{PlotDHLinear}
\title{Linear regression and plotting}
\usage{
PlotDHLinear(pepnum, pepIDS = peptide.identifications, HDXcents, DXcents,
  times)
}
\description{
Calculates linear regression coefficients for HDX and DHX
uptakes, plots the curves and returns the regression coefficients for
the two curves
}
\examples{
coefs <- PlotDHLinear()
}
\author{
Sam MacIntyre \email{smacintyre@csiro.au}
}
\keyword{linear,}
\keyword{regression,}
\keyword{uptakes}
