% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uptake_Plots.R
\name{PlotDHLog}
\alias{PlotDHLog}
\title{Logarithmic regression and plotting}
\usage{
PlotDHLog(pepnum, pepIDS = peptide.identifications, HDXcents, DXcents, times)
}
\description{
Calculates logarithmic regression coefficients for HDX
 and DHX
uptakes, plots the curves and returns the regression coefficients for
the two curves - uses equation of the form: (a*log(x) + b)
}
\examples{
coefs <- PlotDHLog()
}
\author{
Sam MacIntyre \email{smacintyre@csiro.au}
}
\keyword{polynomial,}
\keyword{regression,}
\keyword{uptakes}
