% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uptake_Plots.R
\name{PlotDHPoly2}
\alias{PlotDHPoly2}
\title{Polynomial (order 2) regression and plotting}
\usage{
PlotDHPoly2(pepnum, pepIDS = peptide.identifications, HDXcents, DXcents,
  times)
}
\description{
Calculates polynomial (order 2) regression coefficients for HDX
 and DHX uptakes, plots the curves and returns the regression coefficients for
the two curves
}
\examples{
coefs <- PlotDHPoly2()
}
\author{
Sam MacIntyre \email{smacintyre@csiro.au}
}
\keyword{polynomial,}
\keyword{regression,}
\keyword{uptakes}
