% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_functions.R
\name{plotLin}
\alias{plotLin}
\title{Linear Interpolation Plotting Function}
\usage{
plotLin(interp, peaks)
}
\arguments{
\item{interp}{vector of points of linear interpolation}

\item{peaks}{list of mass peaks}
}
\description{
Given the linear interpolation data and the mass peaks,
plots both over the spectrum
}
\examples{
plotLin(interp,peaks)
}
\author{
Sam MacIntyre \email{smacintyre@csiro.au}
}
\keyword{linear,interpolation,}
\keyword{peaks}
\keyword{plot,}
