% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_functions.R
\name{plotWidth}
\alias{plotWidth}
\title{Width plotting function}
\usage{
plotWidth(v)
}
\arguments{
\item{v}{vector containing qidth}
}
\description{
Given the width vector, plots a horizontal width line 
across the current spectrum
}
\author{
Sam MacIntyre \email{smacintyre@csiro.au}
}
\keyword{plot,}
\keyword{segment}
\keyword{width,}
