% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uptake_Plots.R
\name{PolyIntersect}
\alias{PolyIntersect}
\title{Polynomial intersection}
\usage{
PolyIntersect(PolyCoefs)
}
\description{
Calculates the intersection point of two polynomial functions
given their coefficients
}
\examples{
[x,y] <- PolyIntersect(PolyCoefs)
}
\author{
Sam MacIntyre \email{smacintyre@csiro.au}
}
\keyword{intersection}
\keyword{log,logarithmic,}
