% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculation_functions.R
\name{RecursAppend}
\alias{RecursAppend}
\title{Recursive Appending Function}
\usage{
RecursAppend(list, n)
}
\arguments{
\item{list}{any list}

\item{n}{number of times to replicate}
}
\description{
Given a list, recursively replicates the list for a given
n
}
\examples{
newlist <- RecursAppend(list,4)
}
\author{
Sam MacIntyre \email{smacintyre@csiro.au}
}
\keyword{append,}
\keyword{recursive}
\keyword{replicate,}
