% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculation_functions.R
\name{widthFinder}
\alias{widthFinder}
\title{Peak width calculation}
\usage{
widthFinder(peaks, spectrum, bpint = 0.2)
}
\arguments{
\item{peaks}{List of mass peaks for the spectrum}

\item{spectrum}{A singular mass spectrum}

\item{bpint}{base peak intensity (to determine width)}
}
\description{
Uses the spectrum peaks and a specified base peak intensity
to find the corresponding m/z end points
}
\examples{
[a,b,envelope] <- widthFinder(peaks,spectrum,bpint = 0.5)
}
\author{
Sam MacIntyre \email{smacintyre@csiro.au}
}
\keyword{interpolation}
\keyword{peaks,}
\keyword{width,}
